refactor pages, steps, tests

+++ src/main/java/by/itacademy/moiseenkolydia/breezy/ui/driver/Driver.java
@@ -7,16 +7,18 @@ import org.openqa.selenium.chrome.ChromeOptions;
 public class Driver {
     public static WebDriver driver;
 
-    public static WebDriver warmUp() {
-        ChromeOptions options = new ChromeOptions();
-        options.addArguments("--remote-allow-origins=*", "--incognito", "--lang=en");
-        driver = new ChromeDriver(options);
-        driver.manage().window().maximize();
-        driver.manage().deleteAllCookies();
+    public static WebDriver createDriver() {
+        if (driver == null) {
+            ChromeOptions options = new ChromeOptions();
+            options.addArguments("--remote-allow-origins=*", "--incognito", "--lang=en");
+            driver = new ChromeDriver(options);
+            driver.manage().window().maximize();
+            driver.manage().deleteAllCookies();
+        }
         return driver;
     }
 
-    public void tearDown() {
+    public static void closeDriver() {
         driver.quit();
         driver = null;
     }
